+++ scripts/dataset_prep/prep_v2_format_datasets.py
@@ -1,4 +1,4 @@
-from dataset_service.dataset_defs import AlexandriaDataset
+from dataset_service.datasets import AlexandriaDataset
 from shared import root_dir
 
 AlexandriaDataset().raw_data_to_lmdb(dataset_dir=f"{root_dir}/datasets/alexandria", output_dir=f"{root_dir}/datasets/lmdb/alexandria2")
